REFACTOR Implemented the Factory pattern to encapsulate the creation of responses
By programming to an interface instead of an implementation I can decrease how tricky it will be to change or edit the responses. The interface in this case is the combination of the Factories' method name, the parameter(s) given to it, and its return type - otherwise known as its signature.